miden_utils_indexing

Struct IndexVec

pub struct IndexVec<I: Idx, T> { /* private fields */ }

A dense vector indexed by ID types.

This provides O(1) access and storage for dense ID-indexed data.

Implementations

impl<I: Idx, T> IndexVec<I, T>

pub fn new() -> Self

Create a new empty IndexVec.

pub fn with_capacity(n: usize) -> Self

Create a new IndexVec with pre-allocated capacity.

pub fn len(&self) -> usize

Get the number of elements in the IndexVec.

pub fn is_empty(&self) -> bool

Check if the IndexVec is empty.

pub fn push(&mut self, v: T) -> Result<I, IndexedVecError>

Push an element and return its ID.

Returns an error if the length would exceed the maximum representable by the ID type.

pub fn get(&self, idx: I) -> Option<&T>

Get an element by ID, returning None if the ID is out of bounds.

pub fn as_slice(&self) -> &[T]

Get a slice of all elements.

pub fn into_inner(self) -> Vec<T>

Consume this IndexVec and return the underlying Vec.

pub fn swap_remove(&mut self, index: usize) -> T

Remove an element at the specified index and return it.

pub fn contains(&self, item: &T) -> bool
where T: PartialEq,

Check if this IndexVec contains a specific element.

pub fn iter(&self) -> Iter<'_, T>

Get an iterator over the elements in this IndexVec.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Get a mutable iterator over the elements in this IndexVec.